Enhance the non-200 response from subscriptions
The platform gateway will soon be expecting a specifically formatted response
for dependency errors. This will help more easily triage where entitlement
request failures are originating from, so that response time for remediation
is reduced.

The gateway will implement metrics and alerting based on these responses.

This new contract should be returned by the entitlements service when:
- a non-200 response is received from subscriptions
- an error is received from subscriptions

Both of these scenarios should provide context on the failure, through the
following:

```json
{
  "error": {
    "dependency_failure": true,
    "service": <DEPENDENT_SERVICE_THAT_FAILED>,
    "status": <STATUS_FROM_DEPENDENT_SERVICE>,
    "endpoint": <DEPENDENT_ENDPOINT_THAT_FAILED>,
    "message": <ERROR_MESSAGE_FROM_YOUR_SERVICE>
  }
}
```

This commit adds structs for `DependencyErrorDetails` and `DependencyErrorResponse`,
which are used to supply information during a non-200 failure from the
subscriptions service.