First pass at connecting to cloudwatch
Adds CW config to existing config object, which is then consumed in the logger
upon init.

We need to import the aws and cloudwatch packages into th logger to then connect
to aws and cloudwatch to create the log hook.

This hook will only be created if the cw key exists and we are not in the test
environment.

We should also return early and not attempt to create the hook if there's a
connection issue with cloudwatch.