Support `is_trial` attribute in SKU config
We ultimately need to send a new attribute in the entitlements payload to distinguish
which entitlements have been entitled based on trial SKUs. These SKUs need to be
configured as trial SKUs in our config, by the application bundle owners.

This changes the format of the yaml config from:
```
 - name: ansible
   skus:
     - MCT3691
     - MCT3691RN
```
to
```
 - name: ansible
   skus:
     - MCT3691 {is_trial: false}
     - MCT3691RN {is_trial: true}
```

In response to this change, we need to update the code to unmarshal and consume
the config differently. We do this by introducing a new `SkuAttributes` type,
as well as a new `Skus` type, which is a map of `SkuAttributes` with dynamic keys.

The `Skus` in our `Bundle` type are also updated to be a collection of `Skus`. In
our logic to build the skus for the request, and sku for each bundle to check the
response, we now have to map the keys of our skus config.

I've also added a benchmark into our controller tests for a bundle check. I wanted
to ensure we weren't introducing a regression in performance. The benchmark results
on the test are ~ 2,300 nanoseconds per run on 100k runs. The previous code running
the same benchmark is also ~ 2,300 nanoseconds per run on 100k total, with minor
variance on both.

You can run the benchmarks with the test suite by running `go test -bench=.`