Ensure that we don't set `is_trial: true` when there are no SKUs
In the case where a user has no SKUs (and no trial SKUs), the `len()` check against
the number of common and trial SKUs would evaluate to `true`, since they would
both be 0.

This was causing bundles which the user was not entitled to, to always return
true for `is_trial`.

This fixes that by only doing the length comparison if there is a positive number
of trial SKUs. Otherwise, `is_trial` should remain `false`.