Use new it features api (#55)
* Use new `/featureStatus/` API from IT/subs

**Epic:** https://issues.redhat.com/browse/RHCLOUD-9584

**Background:**
For managing entitlements, we currently have a ceiling of ~850 SKUs which we can
supply to the subscriptions service before we start receiving 400s.

Currently we are manually collecting all SKUs we care about, and passing them to
the subscriptions service: https://github.com/RedHatInsights/entitlements-api-go/blob/4aeef647cabda3341889661f2d487c751fbc239a/controllers/subscriptions.go#L67

Now that IT has provided a `/features` and `/featureStatus` endpoint on the subscriptions
service, we should integrate with it in order to have IT be the source of truth
for telling us whether or not a principal is entitled to a particular feature/bundle
based on their subscriptions, and if that entitlement is a trial/eval or not.

We will still be the source of truth for defining which SKUs each feature/bundle is
represented by, and will push those changes from our config repo into their API.
This will come in the future via: https://issues.redhat.com/browse/RHCLOUD-12009

For now, our bundles are already seeded in all environments based on the current
state: https://github.com/RedHatInsights/entitlements-config/tree/640579dffd0aea66fa970d056627a18f7cd8a178

**Note:** If this state changes prior to automation, we'll need to manually
update the `/features/ansible` or `/features/smart_management` feature in IT
to reflect the latest, so that when we call `/featureStatus/?features=ansible&features=smart_management`
we ensure IT is checking the most recent SKU changes.

**Changes:**
In order to specify the features for the query params to the new endpoint, we've
exposed `ENT_FEATURES` which can be a comma-separated string of values. By default,
this will be set to `ansible,smart_management` but can be configured in the above
environment variable as well.

The `org_id` from the identity header will again be used, so a request will look
something like:
```
https://subscription.api.redhat.com/svcrest/subscription/v5/featureStatus?accountId=<ACCOUNT_ID>&features=ansible&features=smart_management
```

When we get a response back, we'll unmarshal the body into a new `FeatureStatus`
struct. Instead of looping through SKUs from the subs service to determine if a
principal's subs match any SKU in a bundle, we'll let IT handle that and reply
back. Now, we'll get a response similar to:
```
{
  "features": [
    {
      "name": "ansible",
      "isEval": false,
      "entitled": true
    },
    {
      "name": "smart_management",
      "isEval": false,
      "entitled": true
    }
  ]
}
```
We'll use the names of the features (filtered by our query) to match against our
bundles, and when we find a match, set the `is_trial` and `is_entitled` flags
based on that response.

One change is that for these bundle SKUs, we'll default the `is_entitled` flag
to `false` to deny by default.

**Testing:**
This should be easily testable, similar to the old implementation, as the API contract
for this service will not change, so it should be a 1-1 swap. You can verify what
the features have in them by using one of the following `GET` request endpoints (with certs):
```
https://subscription.qa.api.redhat.com/svcrest/subscription/v5/features/ansible
https://subscription.qa.api.redhat.com/svcrest/subscription/v5/features/smart_management
https://subscription.qa.api.redhat.com/svcrest/subscription/v5/features
```

**Future Changes/Considerations:**
- We'll still need to use our bundle files for other non-SKU bundles local to the
  service (and for determining which bundles to check IT features for)
- Maybe we should update the SKU-based bundles to have a "use_feature_status" flag,
  so we can use that vs checking SKUS > 0 in our logic.
- We'll want to pull out the `is_trial` flag and reformat the SKU-based bundles
  since we no longer need that. It's non-breaking as is, however.

* Remove superfluous SKU logic

* Remove superfuous variables