Add prometheus metrics
This adds general, default prometheus metrics which can be access on the `/metrics`
endpoint.

It also adds an explicit counter metric when the service fails to call the subscriptions
service in IT.

As a follow-up to this PR going in, we'll need to ensure prometheus in all environments
is able to scrape our `/metrics` endpoint.

We'll then need to create an alert based on this metric.