Upsert features
Since the subscription features API for POSTing a feature by name is considered
to be an upsert of sorts (technically it will any SKUs, not just additive), we
can make use of this by catching 404s on newly added features, to have the `bundle_synce`
package create these new features for us, based on our `bundles.yml`.

When a feature is added to our config, and is not found when we call `getCurrent`
to retrieve the SKU list, we'll return an empty `SubModel` without error, which
will in turn allow `getUpdates` to collect all the SKUs from `bundles.yml`. They
will all be new, and the POST in `postUpdates` will create the new feature with
the SKUs provided.

Subsequent calls will find the feature, and behave as any other existing feature
by reconciling any differences if needed.