Conditionally run `bundle-sync` on init
Since we have environemnts (FedRAMP/OCM/ephemeral) which share some of the stage
and prod IT endpoints, this means that there is the potential for these environments
to step on the stage/prod features by running the init container if the bundle
config is not in sync with what's in stage/prod.

This will allow us to conditionally, and explicitly enable running the bundle
sync in stage and prod only.