feature: enable automatic certificate renewal for Entitlements
In order to be able to make use of the automatic certificate renewals,
we need to adjust some environment variables that Entitlements can take.
And so as to avoid having downtimes in case we face any issues with it,
the change comes behind a flag which can be enabled or disabled.

RHCLOUD-40635